Partie 2 suppression de view Ajout d'exceptions Reecriture du controleur modification du modele

+++ site/src/main/java/fr/lernejo/travelsite/models/Traveler.java
@@ -1,5 +1,5 @@
 package fr.lernejo.travelsite.models;
 
-public record Traveler(String userEmail, String userName, String userCountry, WeatherExpectation weatherExpectation,float minimumTemperatureDistance) {
+public record Traveler(String userEmail, String userName, String userCountry, WeatherExpectation weatherExpectation,double minimumTemperatureDistance) {
 
 }
